Support multi-byte values with Bitmap
It also corrects the behavior of single byte values.

Fixes #1744

+++ tools/gen_display_resources.py
@@ -153,7 +153,7 @@ const displayio_bitmap_t supervisor_terminal_font_bitmap = {{
     .base = {{.type = &displayio_bitmap_type }},
     .width = {},
     .height = {},
-    .data = (uint32_t*) font_bitmap_data,
+    .data = (size_t*) font_bitmap_data,
     .stride = {},
     .bits_per_value = 1,
     .x_shift = 5,
